+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 
 &nbsp;
 
+# 0.0.11 - 2018-09-15 α
+
+- [x] Added Forward Requests Page under /administrator/transactions/forwardrequests
+- [x] Fixed Ongoing Job Orders Header 
+- [x] Removed dismissable icon in tugboat list in team assignment
+
 # 0.0.10 - 2018-09-13 α
 
 - [x] Added Forward Requests
@@ -26,7 +32,7 @@
 
 `fixes`
 
-### Scripts and Styleshseets are now properly arranged
+### Scripts and Stylesheets are now properly arranged
 
 - scripts now use the ('scripted') section
 - styles now use the ('assets') section